update depedency because of security vulnarability
CVE-2019-11777
Vulnerable versions: < 1.2.1
Patched version: 1.2.1

In the Eclipse Paho Java client library version 1.2.0,
when connecting to an MQTT server using TLS
and setting a host name verifier, the result of that
verification is not checked.
This could allow one MQTT server to impersonate another
and provide the client library with incorrect information.